Initial infrastructure work for methods with Callables
Since we need to be able to call the correct phasers at the right
time. specifically FIRST and LAST, extra logic needs to be added
to find out whether a batch is the last (whether it is the first
can already be deduced from the fact that the ordinal number of
the batch is 0).

The map(Callable) and grep(Callable) candidates have been adapted
to use the slow version of the batcher that provides that info.