Make user dynamic variables visible from hypered code
This is really a terrible hack into the slightly deficient internal
handling of dynamic variables of rakudo in hypered code.

Since we're not interested in any special dynamic variables inside
the ParaSeq code itself, we just assume the first outer context that
we see, for all inner async code execution.